Fix recursive error loop, add OTA keys, add watchdog (currently disabled)
