+++ firmware/esp/README.md
@@ -10,7 +10,7 @@
 * `idfx flash COM2`
 * `idfx monitor COM2`
 
-## Killing a roque OpenOCD app (in case debugger won't let you bind to a port because it's already in use)
+## Killing a rogue OpenOCD app (in case debugger won't let you bind to a port because it's already in use)
 
 `lsof -n -i | grep ":6666"` (or whatever the port is).
