+++ firmware/pico/application/comms/at/cpp_at_settings.hh
@@ -1,0 +1,10 @@
+#ifndef CPP_AT_SETTINGS_HH_
+#define CPP_AT_SETTINGS_HH_
+
+// Override these values to suit your needs!
+#define CPP_AT_COMMAND_MAX_LEN  32
+#define CPP_AT_ARG_MAX_LEN      128
+#define CPP_AT_HELP_STR_MAX_LEN 300
+#define CPP_AT_MAX_NUM_ARGS     10
+
+#endif
