+++ firmware/pico/pico.code-workspace
@@ -92,7 +92,8 @@
 			"shared_mutex": "cpp",
 			"format": "cpp",
 			"esp_loader.h": "c",
-			"comms.hh": "c"
+			"comms.hh": "c",
+			"ads_bee.hh": "c"
 		},
 		"cortex-debug.variableUseNaturalFormat": true,
 		// "cortex-debug.variableUseNaturalFormat": true,
